Remove unnecessary files

+++ setup.sh
@@ -5,14 +5,18 @@ sudo add-apt-repository ppa:linuxuprising/guake
 curl -sL https://deb.nodesource.com/setup_14.x | sudo -E bash -
 sudo apt-get update && apt-get install nodejs zsh curl git guake -y
 
-# zsh
+## zsh
 echo "<><><><><> <><><><><> <><><><><> Installing ZSH..."
 sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
-echo "" > ~/.bash_aliases
+if [ -f ~/.bash_aliases ]; then
+    echo "aliases already initialised"
+else
+    echo "" > ~/.bash_aliases
+fi
 echo "<><><><><> <><><><><> <><><><><> ZSH installed!"
 
-# docker
+## docker
 echo "<><><><><> <><><><><> <><><><><> Installing DOCKER..."
 sudo apt-get install apt-transport-https ca-certificates curl gnupg-agent software-properties-common
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
@@ -26,7 +30,7 @@ sudo groupadd docker
 sudo usermod -aG docker $USER
 echo "<><><><><> <><><><><> <><><><><> DOCKER installed!"
 
-# with CP
+## with CP
 cp -rf ${BASEDIR}/zsh/oh-my-zsh ~/.oh-my-zsh
 cp -rf ${BASEDIR}/zsh/.p10k.zsh ~/.p10k.zsh
 cp -rf ${BASEDIR}/zsh/.bashrc ~/.bashrc
